Updated on 18/12

+++ src/main/java/frc/robot/Constants.java
@@ -10,8 +10,8 @@ public final class Constants {
     public final static class Motors { 
         // Mấy cái này cũng sẽ phải đặt lại theo ID mỗi cái Talon trên Driver Station
         public static int rightMaster = 1;
-        public static int leftMaster = 2;
-        public static int rightFollow = 3;
+        public static int leftMaster = 2; 
+        public static int rightFollow = 3; 
         public static int leftFollow = 4;
 
         public static int intaker1 = 7;
@@ -22,7 +22,7 @@ public final class Constants {
     public final static class Controllers {
         public static int movementController = 0; 
         public static int intakeController = 2; 
-        public static int sensitivity = 1; // 0 -> 1
+        public static double sensitivity = 1; // 0 -> 1
         public static int deadzone = 0; // 0 -> 1
     }
